In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	unsigned long int term1 = 1;
 	unsigned long int term2 = 2;
-	int next;
+	unsigned long int next;
 	int i;
 
 	printf("%lu, ", term1);
